Lots of progress

+++ RTClib/RTClib.cpp
@@ -83,7 +83,7 @@ DateTime::DateTime (uint32_t t) {
     uint8_t leap;
     for (yOff = 0; ; ++yOff) {
         leap = yOff % 4 == 0;
-        if (days < 365 + leap)
+        if (days < (uint16_t)365 + leap)
             break;
         days -= 365 + leap;
     }
